+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.1] - 2022-03-29
+
+### Changed
+- README - update configuration parameters description
+- [cdk-stacks] - Configuration Script (configure.js) - SSM Parameters Put/Get/Delete batching
+- [cdk-stacks] - update dependencies
+- [agent-app] - update dependencies
+- [demo-website] - update dependencies
+
 ## [2.2.0] - 2022-01-06
 
 ### Changed